requirements_n.txt
streamlit
groq
sentence-transformers
faiss-cpu
numpy
pandas
PyPDF2
openpyxl
scikit-learn
torch
